File Seperation
Adds a new Discord collection to store the commands in separate files, so we can clear up the main server.js file. For now, has files for the hello and the jokes command. Basically took the code under the message listener and pasted it into a new file, after setting it up at the beginning of the server.js file (setting up 'fs' or file system, Discord Collection, etc.).